LayoutTests/http/tests/site-isolation/iframe-opener-is-self.html

<!-- webkit-test-runner [ SiteIsolationEnabled=true ] -->
<script>
	if (window.testRunner) {
		window.testRunner.waitUntilDone();
	}
	
	let checkingSame = false;
	window.addEventListener("message", event => {
		if (event.data == "set as self done") {
			window.frames[0].postMessage("check", "*")
			checkingSame = true;
		} else {
			if (checkingSame) {
				console.log(`Cross-origin check: ${event.data}`);
				window.frames[1].postMessage("check", "*")
				checkingSame = false;
			} else {
				console.log(`Same-origin check: ${event.data}`);
				window.testRunner?.dumpAsText();
				window.testRunner?.notifyDone();
			}
		}
	}, false);

	window.onload = () => {
		window.frames[0].postMessage("set as self", "*");
	}
</script>

<iframe src="http://localhost:8000/site-isolation/resources/iframe-opener-is-self-first.html" name="namedframe"></iframe>
<iframe src="http://127.0.0.1:8000/site-isolation/resources/iframe-opener-is-self-second.html"></iframe>
