LayoutTests/http/tests/site-isolation/iframe-opener-is-self-expected.txt

CONSOLE MESSAGE: Cross-origin check: true
CONSOLE MESSAGE: Same-origin check: true
